Разделил код написанный в main.js на отдельные модули
